refactor(ops): Promote ops as a parent mod
