+++ src/ops/bytes_range.rs
@@ -12,407 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-//! Operations used by [`Accessor`][crate::Accessor].
-//!
-//! Users should not use struct or functions here, use [`Operator`][crate::Operator] instead
-
-use std::collections::Bound;
-use std::fmt::Display;
-use std::fmt::Formatter;
 use std::io::Error;
 use std::io::ErrorKind;
 use std::io::Result;
+use std::ops::Bound;
 use std::ops::Range;
 use std::ops::RangeBounds;
 
 use anyhow::anyhow;
-use time::Duration;
-
-use crate::error::other;
-use crate::error::ObjectError;
-use crate::ObjectMode;
-
-/// Operation is the name for APIs in `Accessor`.
-#[derive(Debug, Copy, Clone)]
-pub enum Operation {
-    /// Operation for [`crate::Accessor::metadata`]
-    Metadata,
-    /// Operation for [`crate::Accessor::create`]
-    Create,
-    /// Operation for [`crate::Accessor::read`]
-    Read,
-    /// Operation for [`crate::Accessor::write`]
-    Write,
-    /// Operation for [`crate::Accessor::stat`]
-    Stat,
-    /// Operation for [`crate::Accessor::delete`]
-    Delete,
-    /// Operation for [`crate::Accessor::list`]
-    List,
-    /// Operation for [`crate::Accessor::presign`]
-    Presign,
-}
-
-impl Operation {
-    /// Convert self into static str.
-    pub fn into_static(self) -> &'static str {
-        self.into()
-    }
-}
-
-impl Default for Operation {
-    fn default() -> Self {
-        Operation::Metadata
-    }
-}
-
-impl Display for Operation {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Operation::Metadata => write!(f, "metadata"),
-            Operation::Create => write!(f, "create"),
-            Operation::Read => write!(f, "read"),
-            Operation::Write => write!(f, "write"),
-            Operation::Stat => write!(f, "stat"),
-            Operation::Delete => write!(f, "delete"),
-            Operation::List => write!(f, "list"),
-            Operation::Presign => write!(f, "presign"),
-        }
-    }
-}
-
-impl From<Operation> for &'static str {
-    fn from(v: Operation) -> &'static str {
-        match v {
-            Operation::Metadata => "metadata",
-            Operation::Create => "create",
-            Operation::Read => "read",
-            Operation::Write => "write",
-            Operation::Stat => "stat",
-            Operation::Delete => "delete",
-            Operation::List => "list",
-            Operation::Presign => "presign",
-        }
-    }
-}
-
-/// Args for `create` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpCreate {
-    path: String,
-    mode: ObjectMode,
-}
-
-impl OpCreate {
-    /// Create a new `OpCreate`.
-    ///
-    /// If input path is not match with object mode, an error will be returned.
-    pub fn new(path: &str, mode: ObjectMode) -> Result<Self> {
-        match mode {
-            ObjectMode::FILE => {
-                if path.ends_with('/') {
-                    return Err(other(ObjectError::new(
-                        "create",
-                        path,
-                        anyhow!("Is a directory"),
-                    )));
-                }
-                Ok(Self {
-                    path: path.to_string(),
-                    mode,
-                })
-            }
-            ObjectMode::DIR => {
-                if !path.ends_with('/') {
-                    return Err(other(ObjectError::new(
-                        "create",
-                        path,
-                        anyhow!("Not a directory"),
-                    )));
-                }
-
-                Ok(Self {
-                    path: path.to_string(),
-                    mode,
-                })
-            }
-            ObjectMode::Unknown => Err(other(ObjectError::new(
-                "create",
-                path,
-                anyhow!("create unknown object mode is not supported"),
-            ))),
-        }
-    }
-
-    /// Get path from option.
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-
-    /// Get object mode from option.
-    pub fn mode(&self) -> ObjectMode {
-        self.mode
-    }
-}
-
-/// Args for `read` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpRead {
-    path: String,
-    offset: Option<u64>,
-    size: Option<u64>,
-}
-
-impl OpRead {
-    /// Create a new `OpRead`.
-    ///
-    /// If input path is not a file path, an error will be returned.
-    pub fn new(path: &str, range: impl RangeBounds<u64>) -> Result<Self> {
-        if path.ends_with('/') {
-            return Err(other(ObjectError::new(
-                "read",
-                path,
-                anyhow!("Is a directory"),
-            )));
-        }
-
-        let br = BytesRange::from(range);
-
-        Ok(Self {
-            path: path.to_string(),
-            offset: br.offset(),
-            size: br.size(),
-        })
-    }
-
-    pub(crate) fn new_with_offset(
-        path: &str,
-        offset: Option<u64>,
-        size: Option<u64>,
-    ) -> Result<Self> {
-        if path.ends_with('/') {
-            return Err(other(ObjectError::new(
-                "read",
-                path,
-                anyhow!("Is a directory"),
-            )));
-        }
-
-        Ok(Self {
-            path: path.to_string(),
-            offset,
-            size,
-        })
-    }
-
-    /// Get path from option.
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-
-    /// Get offset from option.
-    pub fn offset(&self) -> Option<u64> {
-        self.offset
-    }
-
-    /// Get size from option.
-    pub fn size(&self) -> Option<u64> {
-        self.size
-    }
-}
-
-/// Args for `stat` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpStat {
-    path: String,
-}
-
-impl OpStat {
-    /// Create a new `OpStat`.
-    pub fn new(path: &str) -> Result<Self> {
-        Ok(Self {
-            path: path.to_string(),
-        })
-    }
-
-    /// Get path from option.
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-}
-
-/// Args for `write` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpWrite {
-    path: String,
-    size: u64,
-}
-
-impl OpWrite {
-    /// Create a new `OpWrite`.
-    ///
-    /// If input path is not a file path, an error will be returned.
-    pub fn new(path: &str, size: u64) -> Result<Self> {
-        if path.ends_with('/') {
-            return Err(other(ObjectError::new(
-                "write",
-                path,
-                anyhow!("Is a directory"),
-            )));
-        }
-
-        Ok(Self {
-            path: path.to_string(),
-            size,
-        })
-    }
-
-    /// Get path from option.
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-
-    /// Get size from option.
-    pub fn size(&self) -> u64 {
-        self.size
-    }
-}
-
-/// Args for `delete` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpDelete {
-    path: String,
-}
-
-impl OpDelete {
-    /// Create a new `OpDelete`.
-    pub fn new(path: &str) -> Result<Self> {
-        Ok(Self {
-            path: path.to_string(),
-        })
-    }
-
-    /// Get path from option.
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-}
-
-/// Args for `list` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpList {
-    path: String,
-}
-
-impl OpList {
-    /// Create a new `OpList`.
-    ///
-    /// If input path is not a dir path, an error will be returned.
-    pub fn new(path: &str) -> Result<Self> {
-        if !path.ends_with('/') {
-            return Err(other(ObjectError::new(
-                "list",
-                path,
-                anyhow!("Not a directory"),
-            )));
-        }
-
-        Ok(Self {
-            path: path.to_string(),
-        })
-    }
-
-    /// Get path from option.
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-}
-
-/// Args for `presign` operation.
-///
-/// The path must be normalized.
-#[derive(Debug, Clone, Default)]
-pub struct OpPresign {
-    path: String,
-    op: Operation,
-    expire: Duration,
-}
-
-impl OpPresign {
-    /// Create a new `OpPresign`.
-    pub fn new(path: &str, op: Operation, expire: Duration) -> Result<Self> {
-        Ok(Self {
-            path: path.to_string(),
-            op,
-            expire,
-        })
-    }
-
-    /// Get path from op.
-    pub fn path(&self) -> &str {
-        &self.path
-    }
-
-    /// Get operation from op.
-    pub fn operation(&self) -> Operation {
-        self.op
-    }
-
-    /// Get expire from op.
-    pub fn expire(&self) -> Duration {
-        self.expire
-    }
-}
-
-/// PresignedRequest is a presigned request return by `presign`.
-///
-/// # TODO
-///
-/// Add signed headers
-#[derive(Debug, Clone)]
-pub struct PresignedRequest {
-    method: http::Method,
-    uri: http::Uri,
-    headers: http::HeaderMap,
-}
-
-impl PresignedRequest {
-    /// Create a new PresignedRequest
-    pub fn new(method: http::Method, uri: http::Uri, headers: http::HeaderMap) -> Self {
-        Self {
-            method,
-            uri,
-            headers,
-        }
-    }
-
-    /// Return request's method.
-    pub fn method(&self) -> &http::Method {
-        &self.method
-    }
-
-    /// Return request's uri.
-    pub fn uri(&self) -> &http::Uri {
-        &self.uri
-    }
-
-    /// Return request's header.
-    pub fn header(&self) -> &http::HeaderMap {
-        &self.headers
-    }
-}
 
 /// BytesRange(offset, size) carries a range of content.
 ///
