+++ src/ops/op_delete.rs
@@ -1,0 +1,37 @@
+// Copyright 2022 Datafuse Labs.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::io::Result;
+
+/// Args for `delete` operation.
+///
+/// The path must be normalized.
+#[derive(Debug, Clone, Default)]
+pub struct OpDelete {
+    path: String,
+}
+
+impl OpDelete {
+    /// Create a new `OpDelete`.
+    pub fn new(path: &str) -> Result<Self> {
+        Ok(Self {
+            path: path.to_string(),
+        })
+    }
+
+    /// Get path from option.
+    pub fn path(&self) -> &str {
+        &self.path
+    }
+}
